gdt refactoring
