+++ kernel/x86/include/x86/gdt.hpp
@@ -3,6 +3,8 @@
 namespace x86 {
 
 class GDT {
+public:
+	void reload();
 };
 
 }
