+++ kernel/x86/src/gdt.cpp
@@ -1,1 +1,6 @@
 #include <x86/gdt.hpp>
+
+void x86::GDT::reload()
+{
+
+}
